# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.CreateTransferDTO:
    properties:
      comment:
        type: string
      from_warehouse_id:
        type: string
      items:
        items:
          $ref: '#/definitions/domain.TransferItemDTO'
        minItems: 1
        type: array
      to_warehouse_id:
        type: string
    required:
    - from_warehouse_id
    - items
    - to_warehouse_id
    type: object
...
  domain.PnLReport:
    properties:
      by_warehouse:
        items:
          $ref: '#/definitions/domain.WarehousePnL'
        type: array
      total_cogs:
        type: number
      total_items_sold:
        type: integer
      total_profit:
        type: number
      total_revenue:
        type: number
    type: object
  domain.TransferItemDTO:
    properties:
      lot_id:
        type: string
      quantity:
        type: integer
    required:
    - lot_id
    - quantity
    type: object
  domain.TransferItemResponse:
    properties:
      id:
        type: string
      quantity:
        type: integer
      source_lot_id:
        type: string
    type: object
  domain.TransferResponse:
    properties:
      accepted_by:
        type: string
      comment:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      from_warehouse_id:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/domain.TransferItemResponse'
        type: array
      status:
        type: string
      to_warehouse_id:
        type: string
    type: object
...
  domain.WarehousePnL:
    properties:
      cogs:
        type: number
      items_sold:
        type: integer
      profit:
        type: number
      revenue:
        type: number
      warehouse_name:
        type: string
    type: object
...
host: localhost:8083
info:
  contact:
    name: Valentyn Khoroshylov
  description: This is a REST API for managing a tires and rims store/warehouse.
  title: Tires Shop CRM API
  version: "1.0"
paths:
  /admin/exports/inventory:
    get:
      parameters:
      - description: Search by brand or model
        in: query
        name: search
        type: string
      - description: Filter by brand name
        in: query
        name: brand
        type: string
      - description: Filter by type (TIRE, RIM)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: URL of the Google Sheet
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - RoleAuth: []
      summary: Export Inventory to Google Sheets
      tags:
      - exports
  /admin/exports/pnl:
    get:
      parameters:
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Filter by Warehouse ID
        in: query
        name: warehouse_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: URL of the Google Sheet
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - RoleAuth: []
      summary: Export P&L to Google Sheets
      tags:
      - exports
...
  /reports/pnl:
    get:
      consumes:
      - application/json
      description: Calculates total revenue, COGS, and profit based on DONE orders.
      parameters:
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Filter by Warehouse ID
        in: query
        name: warehouse_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PnLReport'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - RoleAuth: []
      summary: Get Profit & Loss Report
      tags:
      - reports
...
  /staff/transfers:
    get:
      description: Get paginated list of transfers.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: page_size
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by source warehouse
        in: query
        name: from_warehouse_id
        type: string
      - description: Filter by destination warehouse
        in: query
        name: to_warehouse_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.TransferResponse'
            type: array
      security:
      - RoleAuth: []
      summary: List Transfers
      tags:
      - transfers
    post:
      consumes:
      - application/json
      description: 'Moves stock from one warehouse to another (Status: IN_TRANSIT).'
      parameters:
      - description: Transfer details
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/domain.CreateTransferDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - RoleAuth: []
      summary: Create Transfer
      tags:
      - transfers
  /staff/transfers/{id}:
    get:
      description: Get transfer by ID.
      parameters:
      - description: Transfer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.TransferResponse'
      security:
      - RoleAuth: []
      summary: Get Transfer Details
      tags:
      - transfers
  /staff/transfers/{id}/accept:
    post:
      description: 'Receives the stock at the destination warehouse (Status: ACCEPTED).'
      parameters:
      - description: Transfer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - RoleAuth: []
      summary: Accept Transfer
      tags:
      - transfers
  /staff/transfers/{id}/cancel:
    post:
      description: Cancels an IN_TRANSIT transfer and refunds stock.
      parameters:
      - description: Transfer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - RoleAuth: []
      summary: Cancel Transfer
      tags:
      - transfers
...
securityDefinitions:
  RoleAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"